Add configuration subsystem.
Adds a whole new subsystem for configuration; configuration is provided
in three levels, in the following order:

1. Defaults
2. Values in /etc/gridtogoserver.conf
3. Values specified as parameters on-start

The -c/--config can specify a different configuration file to load. The
default /etc/gridtogoserver.conf is not loaded in this case.

With this commit, there are 3 options (originally 1, -p/--port):

-c/--config
-d/--database
-p/--port

This gives slightly more configurability and the ability to use
persistent configuration will be very useful for server resets.