Add Foreman stuff.
This is mostly for how I've decided to run it, but if you are using the
mongo backend and have "foreman" installed you can use these files to
start it. The ./client.sh and ./server.sh are just generic shell files
that set up a decent PYTHONPATH before running the respective script,
and don't require mongo or foreman.